backend/astro-oraculo/requirements.txt
fastapi
uvicorn[standard]
pyswisseph
pytz
python-multipart
